Move the ManagementController into the common lib
