+++ catalogue_pipeline/id_minter/src/main/scala/uk/ac/wellcome/platform/idminter/Server.scala
@@ -9,7 +9,7 @@ import com.twitter.finatra.http.filters.{
 }
 import com.twitter.finatra.http.routing.HttpRouter
 import uk.ac.wellcome.finatra.modules._
-import uk.ac.wellcome.platform.idminter.controllers.ManagementController
+import uk.ac.wellcome.finatra.controllers.ManagementController
 import uk.ac.wellcome.platform.idminter.modules._
 
 object ServerMain extends Server
