--- catalogue_pipeline/id_minter/src/main/scala/uk/ac/wellcome/platform/idminter/controllers/ManagementController.scala
@@ -1,13 +1,0 @@
-package uk.ac.wellcome.platform.idminter.controllers
-
-import javax.inject.{Inject, Singleton}
-
-import com.twitter.finagle.http.Request
-import com.twitter.finatra.http.Controller
-
-@Singleton
-class ManagementController @Inject()() extends Controller {
-  get("/management/healthcheck") { request: Request =>
-    response.ok.json(Map("message" -> "ok"))
-  }
-}
