+++ sierra_adapter/sierra_bibs_to_dynamo/src/main/scala/uk/ac/wellcome/platform/sierra_bibs_to_dynamo/Server.scala
@@ -9,7 +9,7 @@ import com.twitter.finatra.http.filters.{
 }
 import com.twitter.finatra.http.routing.HttpRouter
 import uk.ac.wellcome.finatra.modules._
-import uk.ac.wellcome.platform.sierra_bibs_to_dynamo.controllers.ManagementController
+import uk.ac.wellcome.finatra.controllers.ManagementController
 import uk.ac.wellcome.platform.sierra_bibs_to_dynamo.modules.{
   SierraBibsToDynamoModule,
   SierraDynamoConfigModule
