--- sierra_adapter/sierra_item_merger/src/main/scala/uk/ac/wellcome/platform/sierra_item_merger/controllers/ManagementController.scala
@@ -1,13 +1,0 @@
-package uk.ac.wellcome.platform.sierra_item_merger.controllers
-
-import javax.inject.{Inject, Singleton}
-
-import com.twitter.finagle.http.Request
-import com.twitter.finatra.http.Controller
-
-@Singleton
-class ManagementController @Inject()() extends Controller {
-  get("/management/healthcheck") { request: Request =>
-    response.ok.json(Map("message" -> "ok"))
-  }
-}
